rename js/bootstrap.js to js/main.js
