+++ source/js/main.js
@@ -1,3 +1,22 @@
-require(['./bootstrap'], function () {
-  //nothing to do here...see bootstrap.js
+/**
+ * bootstraps angular onto the window.document node
+ * NOTE: the ng-app attribute should not be on the index.html when using ng.bootstrap
+ */
+define([
+  'require',
+  'angular',
+  './app',
+  './routes'
+], function (require, ng) {
+  'use strict';
+
+  /*place operations that need to initialize prior to app start here
+   * using the `run` function on the top-level module
+   */
+
+  require(['domReady!'], function (document) {
+    /* everything is loaded...go! */
+    ng.bootstrap(document, ['app']);
+    ng.resumeBootstrap();
+  });
 });
